feat: crd user

+++ apis/users.py
@@ -4,7 +4,7 @@ from flask_restplus import Namespace, Resource, fields, reqparse
 
 from core.db import insert_user, get_user, get_users, delete_users, backup_db
 from core.resource import CustomResource, response, json_serializer_all_datetime_keys
-from core.utils import execute_command_ssh
+from core.utils import execute_command_ssh, check_if_only_int_numbers_exist
 
 
 api = Namespace('users', description='Users related operations')
@@ -41,22 +41,23 @@ def _create_user(name, email, password, user_type=2):
 
 @api.route('/')
 class Users(CustomResource):
+    
     @api.doc('list_users')
+    @api.response(203, 'User does not exist')
     def get(self):
         '''List all users
         
         NOTE: Only for admin users
         '''        
         users = _get_users()
-        if users is None:
-            res = response(status=203)
-        else:
-            res = response(status=200, result=users)
-        return self.send(res)
+        status = 200 if users else 203
+        return self.send(status=200, result=users)
 
 
     @api.doc('create a new user')
     @api.expect(parser_create)
+    @api.response(400, 'Password match error')
+    @api.response(409, 'Duplicate user name')
     def post(self):
         '''Create an user'''
         args = parser_create.parse_args()
@@ -66,28 +67,32 @@ class Users(CustomResource):
         password_confirm = args["password_confirm"]
         
         if password_confirm != password:
-            return self.send(response(status=400))
+            return self.send(status=400, message="Password and Confirm password have to be same")
         if get_user(name=name):
-            return self.send(response(status=409, message="The given username alreay exists."))
+            return self.send(status=409, message="The given username alreay exists.")
         
         result = _create_user(name, email, password)
         status = 201 if result else 500
-        return self.send(response(status=status))
+        return self.send(status=status)
 
 
     @api.doc('delete_users')
     @api.expect(parser_delete)
+    @api.response(400, 'Check user ids')
     def delete(self):
         '''Delete all users
         
         NOTE: Only for admin users or user owner
         '''
         args = parser_delete.parse_args()
-        result = delete_users(args['ids'])
-        return self.send(response(status=200))
+        ids = args['ids']
+        if check_if_only_int_numbers_exist(ids):
+            result = delete_users(args['ids'])
+            return self.send(status=200)
+        else:
+            return self.send(status=400, message="Check user ids")
 
-
-@api.route('/<id_>')
+@api.route('/<int:id_>')
 @api.param('id_', 'The user identifier')
 @api.response(404, 'User not found')
 class User(CustomResource):
@@ -97,13 +102,13 @@ class User(CustomResource):
            
         user = get_user(id_=id_)
         if user is None:
-            return self.send(response(status=404, result=None)) 
+            return self.send(status=404, result=None) 
         user = json_serializer_all_datetime_keys(user)
-        return self.send(response(status=200, result=user))
+        return self.send(status=200, result=user)
         
     
     @api.doc('delete_user')
     def delete(self, id_):
         '''Delete an user'''
         result = delete_users([id_])
-        return self.send(response(status=200))
+        return self.send(status=200)
